requirements.txt
requests>=2.31
xmltodict>=0.13
fastapi>=0.111
uvicorn[standard]>=0.29
fastkml>=1.0
pyyaml>=6.0
shapely>=2.0    
lxml>=4.9        